\name{Lsun3D}
\alias{Lsun3D}

\docType{data}
\title{
Lsun3D inspired by FCPS
}
\description{
clearly defined clusters, different variances
}
\usage{data("Lsun3D")}

\details{
Size 404, Dimensions 3

Dataset defined discontinuites, where the clusters have different variances.
Three main Clusters, and four Outliers (in Cluster 4)
}

\examples{
data(Lsun3D)
str(Lsun3D)
Cls=Lsun3D$Cls
Data=Lsun3D$Data
}
\keyword{datasets}
\keyword{Lsun3D}
\keyword{FCPS}
