\name{sESOM4BMUs}
\alias{sESOM4BMUs}
\title{
simplified ESOM
}
\description{
internfunction for the simplified ESOM Algorithmus [Thrun/Ultsch, 2020] for fixed BestMatchingUnits
}
\usage{
sESOM4BMUs(BMUs,Data, esom, toroid, CurrentRadius,ComputeInR)
}
\arguments{
\item{BMUs}{[1:Lines,1:Columns], BestMAtchingUnits generated by ProjectedPoints2Grid()}
\item{Data}{[1:n,1:d] array of data: n cases in rows, d variables in columns}
\item{esom}{[1:Lines,1:Columns,1:weights] array of NeuronWeights, see ListAsEsomNeurons()}
\item{toroid}{TRUE/FALSE - topology of points}
\item{CurrentRadius}{number betweeen 1 to x}
\item{ ComputeInR}{=T: Rcode, =F Cpp Code}{number betweeen 1 to x}

}


\details{

Algorithm is described in [Thrun, 2018, p. 48, Listing 5.1].
}
\note{Usually not for seperated usage!}

\value{

 \item{esom}{array [1:Lines,1:Columns,1:d], d is the dimension of the weights, the same as in the ESOM algorithm. modified esomneuros regarding a predefined neighborhood defined by a radius}

}

\author{
Michael Thrun
}
\seealso{
\code{\link{GeneralizedUmatrix}}
}
\references{
[Thrun/Ultsch, 2020]  Thrun, M. C., & Ultsch, A.: Uncovering High-Dimensional Structures of Projections from Dimensionality Reduction Methods, MethodsX, Vol. in press, pp. 101093. doi 10.1016/j.mex.2020.101093, 2020.
}

\keyword{GeneratePswarmVisualization}
\keyword{ESOM}
