\name{Uheights4Data}
\alias{Uheights4Data}
\title{Uheights4Data}
\description{
Uheights4Data
}
\usage{
Uheights4Data(BestMatchingUnits, GeneralizedUmatrix)
}
\arguments{
\item{BestMatchingUnits}{(1:n,1:d) BMKey = BestMatchingUnits[,1)}
\item{GeneralizedUmatrix}{(1:Lines,1:Columns) a GeneralizedUmatrix}
}
\value{
\item{Uheights}{Uheights}
\item{BMLineCoords}{BMLineCoords}
\item{BMColCoords}{BMColCoords}
}
\author{
ALU 2021 in matlab, MCT reimplemented in 
}